refactor(filid): split config-loader into single-fn files
Split filid-config.ts (8 fn), rule-docs-manifest.ts (5 fn), and
utils/zod-sanitize.ts (4 fn) into per-function files under loaders/
and utils/. Function bodies are unchanged; only file boundaries and
import/export paths are rewired. config-loader.ts facade re-exports
the same public surface.

- loaders/ gains 14 single-responsibility files (3 schema/type files +
  11 per-function files).
- utils/ gains 6 new single-function helper files.
- exempt-sanitize.ts and rule-engine.ts updated to import FilidConfig /
  AllowedEntry from loaders/config-schemas.js.
- INTENT.md Structure tree and DETAIL.md schema reference updated.
