+++ packages/filid/src/core/infra/config-loader/config-loader.ts
@@ -1,38 +1,31 @@
 /**
  * @file config-loader.ts
- * @description Facade — re-exports all public symbols from loaders/ organ.
+ * @description Facade — re-exports all public symbols from loaders/ and utils/.
  *
- * filid-config: .filid/config.json I/O and project initialisation.
- * rule-docs-manifest: rule doc sync framework (filid-setup skill only).
+ * loaders/: .filid/config.json I/O, project initialisation, rule docs sync.
+ * utils/: resolve-plugin-root (public), other helpers are internal-only.
  */
 export {
   AllowedEntrySchema,
-  createDefaultConfig,
   FilidConfigSchema,
-  initProject,
-  loadConfig,
-  loadRuleOverrides,
-  resolveLanguage,
-  resolveMaxDepth,
   RuleOverrideSchema,
-  validateConfigPatch,
-  writeConfig,
-} from './loaders/filid-config.js';
+} from './loaders/config-schemas.js';
+export type { AllowedEntry, FilidConfig } from './loaders/config-schemas.js';
 export type {
-  AllowedEntry,
   ConfigPatchIssue,
   ConfigPatchValidation,
-  FilidConfig,
   InitResult,
   LoadConfigResult,
-} from './loaders/filid-config.js';
+} from './loaders/config-types.js';
+export { createDefaultConfig } from './loaders/create-default-config.js';
+export { initProject } from './loaders/init-project.js';
+export { loadConfig } from './loaders/load-config.js';
+export { loadRuleOverrides } from './loaders/load-rule-overrides.js';
+export { resolveLanguage } from './loaders/resolve-language.js';
+export { resolveMaxDepth } from './loaders/resolve-max-depth.js';
+export { validateConfigPatch } from './loaders/validate-config-patch.js';
+export { writeConfig } from './loaders/write-config.js';
 
-export {
-  getRuleDocsStatus,
-  loadRuleDocsManifest,
-  syncRuleDocs,
-} from './loaders/rule-docs-manifest.js';
-export { resolvePluginRoot } from './utils/resolve-plugin-root.js';
 export type {
   RuleDocEntry,
   RuleDocStatusEntry,
@@ -40,4 +33,9 @@ export type {
   RuleDocsManifest,
   RuleDocsStatus,
   SyncRuleDocsOptions,
-} from './loaders/rule-docs-manifest.js';
+} from './loaders/manifest-types.js';
+export { getRuleDocsStatus } from './loaders/get-rule-docs-status.js';
+export { loadRuleDocsManifest } from './loaders/load-rule-docs-manifest.js';
+export { syncRuleDocs } from './loaders/sync-rule-docs.js';
+
+export { resolvePluginRoot } from './utils/resolve-plugin-root.js';
